Moved applycriteria to utils and used in SUMIFS
