+++ src/utils/common.js
@@ -1,4 +1,5 @@
 import * as error from './error.js'
+import * as evalExpression from './criteria-eval.js'
 
 // Arrays
 export function argsToArray(args) {
@@ -420,6 +421,60 @@ export function anyIsString() {
 }
 
 // Misc
+//Filters values from a given range based on multiple criteria.
+//Returns an array containing the values that satisfy all the specified criteria.
+export function applyCriteria() {
+  const args = argsToArray(arguments)
+  const range = parseNumberArray(flatten(args.shift()))
+  if (range instanceof Error) {
+    return range
+  }
+
+  const criterias = args
+  const criteriaLength = criterias.length / 2
+
+  for (let i = 0; i < criteriaLength; i++) {
+    criterias[i * 2] = flatten(criterias[i * 2])
+  }
+
+  let values = []
+
+  for (let i = 0; i < range.length; i++) {
+    let isMetCondition = false
+
+    for (let j = 0; j < criteriaLength; j++) {
+      const valueToTest = criterias[j * 2][i]
+      const criteria = criterias[j * 2 + 1]
+      const isWildcard = criteria === void 0 || criteria === '*'
+      let computedResult = false
+
+      if (isWildcard) {
+        computedResult = true
+      } else {
+        const tokenizedCriteria = evalExpression.parse(criteria + '')
+        const tokens = [evalExpression.createToken(valueToTest, evalExpression.TOKEN_TYPE_LITERAL)].concat(
+          tokenizedCriteria
+        )
+
+        computedResult = evalExpression.compute(tokens)
+      }
+
+      // Criterias are calculated as AND so any `false` breaks the loop as unmeet condition
+      if (!computedResult) {
+        isMetCondition = false
+        break
+      }
+
+      isMetCondition = true
+    }
+
+    if (isMetCondition) {
+      values.push(range[i])
+    }
+  }
+  return values
+}
+
 export function isDefined(arg) {
   return arg !== undefined && arg !== null
 }
